Add start script and catch-all route to server
The script to start the application in production mode has been added to the project with the name `start_fe.sh`. In this script, several environment variables are defined and the application's server is started. Furthermore, a catch-all route has been added to `server.js` to serve the front-end application. This ensures that all requests that do not match other routes will serve the `index.html`.